+++ src/app/components/detail/detail.component.html
