+++ src/app/components/detail/review/review.component.css
