+++ src/app/components/detail/review/review.component.html
@@ -2,7 +2,7 @@
     <div class="user-container col col-md-2 align-items-center">
       <div class="row align-items-center">
         <div class="col-auto">
-          <img class="img-fluid user-image" alt="" src="assets/images/profile.png"/>
+          <img class="img-fluid user-image" alt="" src="../../../../assets/images/profile.png"/>
         </div>
         <div class="col-8 col-md">
           <div class="row row-cols-2 row-cols-md-1 align-items-center">
