+++ src/app/interfaces/contentdetails.ts
