+++ src/app/services/platform.service.ts
